后续还需要联立测试
目前测试DHCP问题不大，正在测试TCP和窗口，，接下来需要测试断开，还需要制造模拟数据以供测试

+++ Z-TCP/Z-TCP/SourceCode/ARP.c
@@ -68,7 +68,7 @@ uint8_t ARP_GetMAC_ByIP(IP * ip, MAC * mac, uint8_t * IndexOfCache, uint8_t Send
 	return ARP_False;
 }
 
-void ARP_AddItem(IP * ip, MAC * mac,uint8_t TTL)
+void ARP_AddItem(IP * ip, MAC * mac,uint16_t TTL)
 {
 	uint8_t IndexOfCache = 0,i;
